+++ src/Firebase/Auth/UserInfo.php
@@ -26,14 +26,6 @@ class UserInfo implements \JsonSerializable
         return $info;
     }
 
-    /**
-     * @deprecated 4.33
-     */
-    public function toArray(): array
-    {
-        return \get_object_vars($this);
-    }
-
     public function jsonSerialize()
     {
         return \get_object_vars($this);