+++ src/Firebase/Exception/Database/DatabaseError.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Exception\Database;
 
 use Kreait\Firebase\Exception\DatabaseException;
-use Kreait\Firebase\Exception\HasRequestAndResponse;
 use RuntimeException;
 
 final class DatabaseError extends RuntimeException implements DatabaseException
 {
-    use HasRequestAndResponse;
 }