+++ src/Firebase/Exception/Database/TransactionFailed.php
@@ -4,17 +4,13 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Exception\Database;
 
-use GuzzleHttp\Exception\RequestException;
 use Kreait\Firebase\Database\Reference;
 use Kreait\Firebase\Exception\FirebaseException;
-use Kreait\Firebase\Exception\HasRequestAndResponse;
 use RuntimeException;
 use Throwable;
 
 final class TransactionFailed extends RuntimeException implements FirebaseException
 {
-    use HasRequestAndResponse;
-
     /** @var Reference */
     private $reference;
 
@@ -35,19 +31,6 @@ final class TransactionFailed extends RuntimeException implements FirebaseExcept
         parent::__construct($message, $code, $previous);
 
         $this->reference = $query;
-
-        if ($previous && $previous instanceof RequestException) {
-            $requestException = $previous;
-        } elseif ($previous && ($prePrevious = $previous->getPrevious()) && ($prePrevious instanceof RequestException)) {
-            $requestException = $prePrevious;
-        } else {
-            $requestException = null;
-        }
-
-        if ($requestException) {
-            $this->request = $requestException->getRequest();
-            $this->response = $requestException->getResponse();
-        }
     }
 
     public static function onReference(Reference $reference, Throwable $error = null): self