+++ src/Firebase/ServiceAccount.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Kreait\Firebase;
 
 use Kreait\Firebase\Exception\InvalidArgumentException;
-use Kreait\Firebase\ServiceAccount\Discoverer;
 use Kreait\Firebase\Util\JSON;
 use Throwable;
 
@@ -17,70 +16,9 @@ class ServiceAccount
     /** @var array */
     private $data = [];
 
-    /** @var string|null */
-    private $filePath;
-
-    /**
-     * @deprecated 4.42.0
-     *
-     * @return string|null
-     */
-    public function getFilePath()
-    {
-        return $this->filePath;
-    }
-
     public function getProjectId(): string
     {
         return $this->data['project_id'] ?? '';
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function getSanitizedProjectId(): string
-    {
-        $sanitized = \preg_replace('/[^A-Za-z0-9\-]/', '-', $projectId = $this->getProjectId());
-
-        return \is_string($sanitized) ? $sanitized : $projectId;
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function withProjectId(string $value): self
-    {
-        $serviceAccount = clone $this;
-        $serviceAccount->data['project_id'] = $value;
-
-        return $serviceAccount;
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function hasClientId(): bool
-    {
-        return $this->getClientId() !== '';
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function getClientId(): string
-    {
-        return $this->data['client_id'] ?? '';
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function withClientId(string $value): self
-    {
-        $serviceAccount = clone $this;
-        $serviceAccount->data['client_id'] = $value;
-
-        return $serviceAccount;
     }
 
     public function getClientEmail(): string
@@ -88,43 +26,9 @@ class ServiceAccount
         return $this->data['client_email'] ?? '';
     }
 
-    /**
-     * @deprecated 4.42.0
-     */
-    public function withClientEmail(string $value): self
-    {
-        if (!\filter_var($value, \FILTER_VALIDATE_EMAIL)) {
-            throw new InvalidArgumentException(\sprintf('"%s" is not a valid email.', $value));
-        }
-
-        $serviceAccount = clone $this;
-        $serviceAccount->data['client_email'] = $value;
-
-        return $serviceAccount;
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function hasPrivateKey(): bool
-    {
-        return $this->getPrivateKey() !== '';
-    }
-
     public function getPrivateKey(): string
     {
         return $this->data['private_key'] ?? '';
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public function withPrivateKey(string $value): self
-    {
-        $serviceAccount = clone $this;
-        $serviceAccount->data['private_key'] = \str_replace('\n', "\n", $value);
-
-        return $serviceAccount;
     }
 
     public function asArray(): array
@@ -171,13 +75,7 @@ class ServiceAccount
         throw new InvalidArgumentException('Invalid service account specification.');
     }
 
-    /**
-     * @internal
-     *
-     * @deprecated 4.42.0 Use {@see \Kreait\Firebase\ServiceAccount::fromValue()} instead.
-     * @see fromValue()
-     */
-    public static function fromArray(array $data): self
+    private static function fromArray(array $data): self
     {
         $type = $data['type'] ?? '';
 
@@ -196,26 +94,14 @@ class ServiceAccount
         return $serviceAccount;
     }
 
-    /**
-     * @internal
-     *
-     * @deprecated 4.42.0 Use {@see \Kreait\Firebase\ServiceAccount::fromValue()} instead.
-     * @see fromValue()
-     */
-    public static function fromJson(string $json): self
+    private static function fromJson(string $json): self
     {
         $config = JSON::decode($json, true);
 
         return self::fromArray($config);
     }
 
-    /**
-     * @internal
-     *
-     * @deprecated 4.42.0 Use {@see \Kreait\Firebase\ServiceAccount::fromValue()} instead.
-     * @see fromValue()
-     */
-    public static function fromJsonFile(string $filePath): self
+    private static function fromJsonFile(string $filePath): self
     {
         try {
             $file = new \SplFileObject($filePath);
@@ -230,32 +116,6 @@ class ServiceAccount
             throw new InvalidArgumentException(\sprintf('%s could not be parsed to a Service Account: %s', $filePath, $e->getMessage()));
         }
 
-        $serviceAccount->filePath = $filePath;
-
         return $serviceAccount;
-    }
-
-    /**
-     * @deprecated 4.42.0
-     *
-     * @internal
-     */
-    public static function withProjectIdAndServiceAccountId(string $projectId, string $serviceAccountId): self
-    {
-        return self::fromArray([
-            'type' => 'service_account',
-            'project_id' => $projectId,
-            'client_email' => $serviceAccountId,
-        ]);
-    }
-
-    /**
-     * @deprecated 4.42.0
-     */
-    public static function discover(Discoverer $discoverer = null): self
-    {
-        $discoverer = $discoverer ?: new Discoverer();
-
-        return $discoverer->discover();
     }
 }