+++ tests/Unit/MessagingTest.php
@@ -121,7 +121,7 @@ class MessagingTest extends UnitTestCase
 
         $this->messagingApi
             ->method('send')
-            ->willThrowException((new NotFound())->withResponse(new Response()));
+            ->willThrowException((new NotFound()));
 
         $this->expectException(InvalidMessage::class);
         $this->messaging->validate($message);